transferencias-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
google-cloud-firestore==2.13.0
firebase-admin==6.3.0
httpx==0.25.2
python-dotenv==1.0.0